norminetted first part and added comments to functions

+++ ft_strlen.c
@@ -6,15 +6,17 @@
 /*   By: rmamzer <rmamzer@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/16 18:59:54 by rmamzer           #+#    #+#             */
-/*   Updated: 2025/04/21 10:28:13 by rmamzer          ###   ########.fr       */
+/*   Updated: 2025/04/28 14:23:10 by rmamzer          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
+// The  strlen() function calculates the length of the string pointed to by s,
+// excluding the terminat‐ing null byte ('\0').
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
